Further refactoring to put classes in the right place.
