+++ src/test/java/org/project_kessel/relations/client/CDIManagedRelationsClientsContainerTests.java
@@ -1,4 +1,4 @@
-package org.project_kessel.clients;
+package org.project_kessel.relations.client;
 
 import org.project_kessel.api.relations.v1beta1.*;
 import io.grpc.Server;
@@ -13,9 +13,7 @@ import org.jboss.weld.junit5.WeldSetup;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
-import org.project_kessel.relations.client.CheckClient;
-import org.project_kessel.relations.client.LookupClient;
-import org.project_kessel.relations.client.RelationTuplesClient;
+import org.project_kessel.clients.Config;
 
 import java.io.IOException;
 import java.util.Optional;
